requirements/base.txt
Django==1.9
djangorestframework==3.3.2
django-versatileimagefield==1.2
django-cors-headers==1.1.0
django-storages==1.1.8
djangorestframework-gis==0.10
django-floppyforms==1.6.0
bootstrap-admin==0.3.6
Werkzeug==0.11.2
Markdown==2.6.5
dj-database-url==0.3.0
dj-static==0.0.6
django-filter==0.11
django-autoslug==1.9.3
django-rest-swagger==0.3.4
Pillow==2.8.1
pytz==2015.7
PyYAML==3.11
six==1.10.0
static3==0.5.1
wsgiref==0.1.2
sqlparse==0.1.18
